Fysbok

+++ server/app/Http/Controllers/Api/AnnonceController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
-use App\Http\Resources\AnnonceCollection;
-use App\Http\Resources\AnnonceResource;
+use App\Models\User;
 use App\Models\Annonce;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use App\Http\Controllers\Controller;
+use App\Http\Resources\AnnonceResource;
+use App\Http\Resources\AnnonceCollection;
 
 class AnnonceController extends Controller
 {
@@ -44,9 +46,20 @@ class AnnonceController extends Controller
      */
     public function store(Request $request)
     {
+        if ($request->email) {
 
-        $annonce = Annonce::create($request->all());
-        return new AnnonceResource($annonce);
+            // find user by email
+            $user = User::where('email', $request->email)->first();
+            $admin_id = DB::table('roles')->where('name', 'admin')->first()->id;
+            if ($user->role_id == $admin_id) {
+                $annonce = Annonce::create($request->all());
+                return new AnnonceResource($annonce);
+            } else {
+                return response("You are not allowed to create an annonce", 401);
+            }
+        }
+        // return json response
+        return response("You are not allowed to create an annonce", 401);
     }
 
     /**
@@ -80,11 +93,17 @@ class AnnonceController extends Controller
      */
     public function update(Request $request, Annonce $annonce)
     {
-
-
-        $annonce->update($request->all());
-        return new AnnonceResource($annonce);
+        $user = User::where('email', $request->email)->first();
+        $admin_id = DB::table('roles')->where('name', 'admin')->first()->id;
+        if ($user->role_id == $admin_id) {
+            $annonce->update($request->all());
+            return new AnnonceResource($annonce);
+        } else {
+            return response("You are not allowed to update an annonce", 401);
+        }
+        return response("You are not allowed to update an annonce", 401);
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -92,10 +111,17 @@ class AnnonceController extends Controller
      * @param  \App\Models\Annonce  $annonce
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Annonce $annonce)
+    public function destroy(Annonce $annonce, Request $request)
     {
+        $user = User::where('email', $request->email)->first();
+        $admin_id = DB::table('roles')->where('name', 'admin')->first()->id;
+        if ($user->role_id == $admin_id) {
 
-        $annonce->delete();
-        return response("Annonce deleted successfully", 204);
+            $annonce->delete();
+            return response("Annonce deleted successfully", 200);
+        } else {
+            return response("You are not allowed to update an annonce", 401);
+        }
+        return response("You are not allowed to update an annonce", 401);
     }
 }
